requirements.txt
streamlit
google-generativeai
pinecone-client
pypdf
python-dotenv
